+++ viewmodel/src/commonTest/kotlin/com/hoc081098/kmp/viewmodel/utils/runBlockInNewThread.kt
@@ -1,3 +1,3 @@
-package com.hoc081098.kmp.viewmodel
+package com.hoc081098.kmp.viewmodel.utils
 
 expect suspend fun runBlockInNewThread(block: () -> Unit)