+++ viewmodel/src/jsTest/kotlin/com/hoc081098/kmp/viewmodel/utils/runBlockInNewThread.js.kt
@@ -1,3 +1,3 @@
-package com.hoc081098.kmp.viewmodel
+package com.hoc081098.kmp.viewmodel.utils
 
 actual suspend fun runBlockInNewThread(block: () -> Unit) = block()